requirements.txt
yara-python==4.3.1